very early WIP to override and expand the hardcoded VEH array of 1028
extract_veh_disp_offsets.py is a jython for Ghidra to emit all ptr refernces to the original veh array.  It still need to be fleshed out...

It outputs to veh_disp_offsets.tsv which is then parsed with generate_veh_patch.py which outputs a mostly "drop in" .cpp

Initial testing with a new game saves and loads fine.
I tested with a "late game" save and it failed to load looping infinitely at 005BFF83 a while loop which I suspect is due to a bad prevVehIDStack ID.